+++ webapp/src/app/components/toolbar/toolbar.component.ts
@@ -32,7 +32,7 @@ export class ToolbarComponent implements OnInit {
 	public loadMapClicked = new EventEmitter<void>(false);
 	
 	constructor(private mapLoader: MapLoaderService,
-		private events: GlobalEventsService,
+		public events: GlobalEventsService,
 		private dialog: MatDialog,
 		private overlayService: OverlayService,
 		private overlay: Overlay,
@@ -123,8 +123,12 @@ export class ToolbarComponent implements OnInit {
 		});
 	}
 	
-	changeTo3d(event: MatSlideToggleChange) {
-		this.is3d = event.checked;
-		this.router.navigate([event.checked ? '3d' : '']);
+	changeTo3d(checked: boolean) {
+		this.is3d = checked;
+		this.router.navigate([checked ? '3d' : '']);
+	}
+	
+	toggleIngamePreview(checked: boolean) {
+		this.events.showIngamePreview.next(checked);
 	}
 }