+++ webapp/src/app/services/phaser/tilemap/cc-map.ts
@@ -144,8 +144,13 @@ export class CCMap {
 		this.mapWidth = width;
 		this.mapHeight = height;
 		
-		this.layers.forEach(layer => layer.resize(width, height));
-		Globals.phaserEventsService.updateMapBorder.next(true);
+		this.layers.forEach(layer => {
+			// only update layers with distance: 1
+			// Parallax Layers shouldn't be updated because they usually have different dimensions than the map
+			if (layer.details.distance === 1) {
+				layer.resize(width, height);
+			}
+		});
 	}
 	
 	offsetMap(offset: Point, borderTiles = false) {